+++ packages/core/src/components/essentials/title.ts
@@ -11,7 +11,7 @@ export class Title extends Component {
 	// eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-	render(animate = true) {
+	render(animate = false) {
 		const svg = this.getComponentContainer()
 		const title = getProperty(this.getOptions(), 'title')