+++ packages/core/vite.config.ts
@@ -1,6 +1,6 @@
 /// <reference types="vitest" />
 import { fileURLToPath } from 'url'
-import { defineConfig } from 'vite'
+import { defineConfig, type PluginOption } from 'vite'
 import dts from 'vite-plugin-dts'
 
 export default defineConfig({
@@ -31,7 +31,7 @@ export default defineConfig({
 			'@carbon/utils-position',
 			'carbon-components',
 			'date-fns', // make peerDependency and externalize in next major
-			'dom-to-image-more',
+			'html-to-image',
 			'lodash-es' // make peerDependency and externalize in next major (or replace with modern TypeScript alternatives)
 		],
 		exclude: [
@@ -44,9 +44,21 @@ export default defineConfig({
 		}
 	},
 	plugins: [
-		// equivalent to: npx tsc src/*.ts --declaration --emitDeclarationOnly --declarationDir dist/ --resolveJsonModule --esModuleInterop
+		// Equivalent to: npx tsc --declaration --emitDeclarationOnly --rootDir src
+		// However, tsc does not translate '@/' correctly so we need to use vite-plugin-dts. TypeScript, on its own, only allows
+		// type, not compilation, error suppression.
 		dts({
-			exclude: './package.json'
-		})
+			/*
+			* Note: src/demo/utils/package-versions.ts imports package.json files from other packages which are outside of the root directory.
+			* This causes vite-plugin-dts to redefine the rootDir which results in type declarations going into dist/dist/core/src.
+			* The function below fixes their path.
+			*/
+			// Leave for next version of vite-plugin-dts
+			// beforeWriteFile: (filePath: string, content: string) => {
+			// 	filePath = filePath.replace('/dist/dist/core/src/','/dist/')
+			// 	return { filePath, content }
+			// },
+			logLevel: 'silent' // Suppress the known errors because package.json files don't need d.ts files
+		}) as unknown as PluginOption
 	]
 })