+++ packages/angular/CHANGELOG.md
@@ -3,6 +3,59 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.6](https://github.com/carbon-design-system/carbon-charts/compare/v1.11.5...v1.11.6) (2023-07-06)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.11.5](https://github.com/carbon-design-system/carbon-charts/compare/v1.11.4...v1.11.5) (2023-07-06)
+
+### Bug Fixes
+
+- **react:** remount issues
+  ([#1606](https://github.com/carbon-design-system/carbon-charts/issues/1606))
+  ([aad4838](https://github.com/carbon-design-system/carbon-charts/commit/aad4838771a398b7e37484c60b4f9fc0060e1a90))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.11.4](https://github.com/carbon-design-system/carbon-charts/compare/v1.11.3...v1.11.4) (2023-07-06)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.11.3](https://github.com/carbon-design-system/carbon-charts/compare/v1.11.2...v1.11.3) (2023-07-03)
+
+### Bug Fixes
+
+- StackBlitz dependencies & render(animate: boolean) values
+  ([#1601](https://github.com/carbon-design-system/carbon-charts/issues/1601))
+  ([40cb8ec](https://github.com/carbon-design-system/carbon-charts/commit/40cb8ecb810cdacd3a9bd89c64b6f1bde3b3a202))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [1.11.2](https://github.com/carbon-design-system/carbon-charts/compare/v1.11.1...v1.11.2) (2023-06-28)
+
+**Note:** Version bump only for package @carbon/charts-angular
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 ## [1.11.1](https://github.com/carbon-design-system/carbon-charts/compare/v1.11.0...v1.11.1) (2023-06-28)
 
 **Note:** Version bump only for package @carbon/charts-angular